requirements.txt
# --------------------- If use pip -----------------------------
Cython
numpy
opencv-python
torch
torchvision
onnx
django
djangorestframework
matplotlib
PyYAML
scipy
tqdm
tensorboard
pillow
thop
scikit-image
# git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI
pycocotools
pathlib
setuptools
google-utils

# for reading files directly from s3
s3fs


# ------------------- Tensorboard ------------------------------
# tb-nightly
# future

# ------------------ If use anconda -----------------------------
# conda update -yn base -c defaults conda
# conda install -yc anaconda numpy opencv matplotlib tqdm pillow ipython future
# conda install -yc conda-forge scikit-image pycocotools tensorboard
# conda install -yc spyder-ide spyder-line-profiler
# conda install -yc pytorch pytorch torchvision
# conda install -yc conda-forge protobuf numpy && pip install onnx==1.6.0  # https://github.com/onnx/onnx#linux-and-macos

